test: #84: add unit tests for DateTimeExtensions methods

+++ Core/Extensions/DateTimeExtensions.cs
@@ -2,7 +2,7 @@ public static class DateTimeExtensions
 {
     public static decimal GetHours(this DateTime startTime, DateTime endTime)
     {
-        return startTime.GetTotalMinutes(endTime) / 60;
+        return (decimal)startTime.GetTotalMinutes(endTime) / 60;
     }
 
     public static int GetTotalMinutes(this DateTime startTime, DateTime endTime)
